Add files via upload
The purpose of uploading the Dockerfile is so that anyone wanting to use the container can see what packages are installed.
The Dockerfile is not required to run the image, which is on dockerhub. The container runs correctly on Windows (using wsl2), Linux, Azure, and AWS.

To run the kf7032 docker image:
docker run -it --rm 8888:8888 -p 8080:8080 -p 8081:8081 -p 4040:4040 -v /home/source_path:/home/notebookuser/working jeremyellman/kf7032_22s_b:latest
This maps the ports used by Apache Spark and a source file path to the container folder /home/notebookuser/working

The container has a password since this is required by internet facing jupyter notebooks. Since the password is saved, users need not use usual jupyter token method.